Implemented Aesthetic Only Player Interface
A new class, PlayerInterface, draws a box and handles a custom cursor
with a click animation. The structure for the drag box will make it easy
to detect if we are selecting units later.

Also, made it that the debug grid dosent show unless you are pushing G,
good for demoing I think.