Improved Nav-Grid API
New Functions :
GetNodeClosestToPixelPos() - Used to get the closest node to the mouse
pos
GetNodeNeighbor() (Untested) - Can get the neighbor of a node directly
NodeIsOnGrid() - Just a quick function that tests if a node is on the
grid or not

All the Nav-Grid stuff now returns pointers now, so you can use the
nodes directly to store pathfinding data.

A little big of debug screen stuff added too, hit G to see.